Prefer pixel precision on iTerm
This change is specific to https://github.com/antonio-ramadas/seam-carving.
I want to be able to control the images' size without much fuss.
Previous dimensions controlled the size of image by character cells
which I found harder to tune. However, iTerm supports pixel precision
which I find easier to reason about.
Because this change is not expected to be merged with [upstream](https://github.com/atanunq/viuer)
 and to be able to hardwire with my other project, this commit also sets
 a specific version distinct enough for it to be evident that something
unusual is going on.